[cli] Re-organize source of truth for return codes
